requirements.txt
colorama==0.4.6
coverage==7.6.12
iniconfig==2.0.0
memory-profiler==0.61.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psutil==7.0.0
pytest==8.3.4
pytest-cov==6.0.0
tkinterdnd2==0.4.2
